[config-plugins] extract build scheme functions to separate module
