+++ packages/config-plugins/src/ios/index.ts
@@ -1,5 +1,6 @@
 import * as AdMob from './AdMob';
 import * as Branch from './Branch';
+import * as BuildScheme from './BuildScheme';
 import * as BundleIdentifier from './BundleIdentifier';
 import * as CustomInfoPlistEntries from './CustomInfoPlistEntries';
 import * as DeviceFamily from './DeviceFamily';
@@ -17,7 +18,6 @@ import * as ProvisioningProfile from './ProvisioningProfile';
 import * as RequiresFullScreen from './RequiresFullScreen';
 import * as Scheme from './Scheme';
 import * as SplashScreen from './SplashScreen';
-import * as Target from './Target';
 import * as Updates from './Updates';
 import * as UserInterfaceStyle from './UserInterfaceStyle';
 import * as UsesNonExemptEncryption from './UsesNonExemptEncryption';
@@ -30,6 +30,7 @@ import * as XcodeUtils from './utils/Xcodeproj';
 export {
   AdMob,
   Branch,
+  BuildScheme,
   BundleIdentifier,
   CustomInfoPlistEntries,
   DeviceFamily,
@@ -48,7 +49,6 @@ export {
   Permissions,
   RequiresFullScreen,
   Scheme,
-  Target,
   Updates,
   UserInterfaceStyle,
   UsesNonExemptEncryption,
